feat(api)!: add pagination to entities list
This change introduces pagination to the `GET /api/entities` endpoint to allow for more efficient handling of large sets of entity data.

The entities service and API endpoint now support `page` and `limit` query parameters. The response has been updated to include a `pagination` object with total results and page information.

BREAKING CHANGE: The response shape for `GET /api/entities` has changed. The root of the response is now an object containing `data` and `pagination` properties, rather than a simple array of entities.